requirements.txt
# MySoulMate Flask Dependencies
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
neo4j==5.15.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
certifi==2023.11.17